Collapse "tool-get-webpage".
